backend/requirements.txt
# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0

# MetaTrader 5
MetaTrader5==5.0.5120


# Análisis técnico y datos
pandas>=2.2.2
numpy==1.24.3
talib-binary==0.4.26
scipy==1.11.4

# Machine Learning e IA
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSockets
websockets==12.0

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging y monitoreo
structlog==23.2.0

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Visualización (para desarrollo)
matplotlib==3.8.2
plotly==5.17.0

# Timezone handling
pytz==2023.3

# Async utilities
asyncio-mqtt==0.13.0

# Data validation
email-validator==2.1.0